test(subprocess): type binary reads and native test spies precisely

+++ packages/sandbox/sandbox-windows-acl/tests/control.spec.ts
@@ -54,7 +54,7 @@ describe.skipIf(process.platform !== 'win32')('managed Windows ACL control pipe'
     for (let index = 0; index < input.length; index++) input[index] = index % 256
     const received = (async () => {
       const chunks: Buffer[] = []
-      for await (const chunk of channel) chunks.push(Buffer.from(chunk))
+      for await (const chunk of channel) chunks.push(Buffer.from(chunk as Uint8Array))
       return Buffer.concat(chunks)
     })()
     channel.write(input)
